Avoid type conflicts with the 4.0 System.Net.Http dll
When running this in Unity with the 4.6 experimental runtime, Unity will load Mono's version of System.Net.Http, which causes type conflicts all over the place given that both dlls are running side by side. Easiest way of avoiding that is to move this version into its own namespace - DotNetHttp35.
Also renamed the dll to reflect the namespace change.